Safari-609.1.11.1/LayoutTests/fast/dom/Element/outerHTML-prefixed-attribute.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<svg xmlns="http://www.w3.org/2000/svg">
  <g></g>
</svg>
<script>
description("Makes sure prefixes are properly serialized by Element.outerHTML in an HTML document.");

let g = document.querySelector("g")
g.setAttributeNS("http://example.com","ex:attr","1")
shouldBeEqualToString("g.outerHTML", "<g ex:attr=\"1\"></g>");
</script>
</body>
</html>
